+++ Aplia/Migration/ContentMigration.php
@@ -1,0 +1,28 @@
+<?php
+namespace Aplia\Migration;
+use Phinx\Migration\AbstractMigration;
+use eZDB;
+
+/**
+ * Defines a migration step which takes care of setting up
+ * eZ publish database transactions.
+ * 
+ * Note: This requires that Phinx is installed.
+ */
+abstract class ContentMigration extends AbstractMigration
+{
+    public function change()
+    {
+        $db = eZDB::instance();
+        $db->begin();
+
+        $this->changeContent();
+
+        $db->commit();
+    }
+
+    /**
+     * Changes eZ publish content, migration class must implement this.
+     */
+    public abstract function changeContent();
+}